ContactsCommon: Display sim contact when set the phone contact to display
In this case, we try filter a specified account when one sim card is
disabled, so we will append different account name to the query parameter,
it will cause the query can not distinguish which acount name is used for
filter the specified account and which is used for disabled sim card's
contacts, so it will cause the display error.

When we filter a specified account, it not necessary to filter the disable
sim card's contacts at same time, so remove the diabled sim filter when we
filter a specified account.

CRs-Fixed: 618811

Change-Id: I5ef18b05c9a6ce676164e3054178f31753cfa40e